+++ src/main/client/system/index.ts
@@ -4,6 +4,7 @@ import './consoleCommand.js';
 import './native.js';
 import './notification.js';
 import './raycasts.js';
+import './serverConfig.js';
 import './stats.js';
 import './waypoint.js';